pyproject.toml
[tool.poetry]
name = "biometric-recognition"
version = "0.1.0"
description = "Multimodal Biometric Recognition System using Fingerprints and Iris"
authors = ["Anirudh <anirudhsridhar1905@gmail.com>"]
packages = [{include = "biometric_recognition", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
# Core ML framework (required for both API and pipeline)
torch = "^2.10.0"
timm = "^0.9.0"
# Data handling
numpy = "^1.24.0"
pillow = "^10.0.0"
# Configuration
omegaconf = "^2.3.0"

[tool.poetry.group.api]
optional = true

[tool.poetry.group.api.dependencies]
# API serving
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
# S3 model loading
boto3 = "^1.34.0"

[tool.poetry.group.pipeline]
optional = true

[tool.poetry.group.pipeline.dependencies]
# Evaluation and visualization
scikit-learn = "^1.3.0"
matplotlib = "^3.7.0"
# Configuration
hydra-core = "^1.3.0"
tqdm = "^4.65.0"
# AWS S3 support
boto3 = "^1.34.0"
# Experiment tracking
mlflow = "^2.20.0"
setuptools = "^82.0.0"
# Fix pyarrow build issue on Python 3.13
pyarrow = ">=17.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
pre-commit = "^3.3.0"
jupyter = "^1.0.0"


[tool.poetry.scripts]
train = "biometric_recognition.train:main"
serve = "biometric_recognition.api.serve:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
